Moved dbstat.h below subfolder eponymous_vtabs
There are many other so-called "eponymous virtual tables", so start organizing them
