+++ dev/eponymous_vtabs/dbstat.h
@@ -2,8 +2,8 @@
 
 #include <string>  //  std::string
 
-#include "column.h"
-#include "table.h"
+#include "../column.h"
+#include "../table.h"
 
 namespace sqlite_orm {
 #ifdef SQLITE_ENABLE_DBSTAT_VTAB
